refactor: move hash to esxi os
