+++ tests/plugins/os/unix/esxi/test__os.py
@@ -4,6 +4,7 @@ from dissect.target.filesystem import VirtualFilesystem
 from dissect.target.plugins.os.unix.esxi._os import (
     _create_local_fs,
     _decrypt_crypto_util,
+    esxi_hash,
 )
 from dissect.target.target import Target
 
@@ -58,3 +59,22 @@ def test__decrypt_crypto_local(fs_unix: VirtualFilesystem) -> None:
         mocked_run.return_value.stdout = b"data"
 
         assert _decrypt_crypto_util(fs_unix.path("data")).read() == b"data"
+
+
+def test_hash_remainder() -> None:
+    ip = b"192.168.1.109"
+    volume = b"/home/roel/nfstest"
+    h1 = esxi_hash(ip, 42)
+    h2 = esxi_hash(volume, h1)
+    assert h2 == 5364432747070711354
+
+
+def test_hash_full() -> None:
+    h1 = esxi_hash(b"Het implementeren van hashfuncties in Python is lastiger dan je zou denken,", 42)
+    h2 = esxi_hash(b"met name door de ontbrekende ondersteuning voor unsigned integer arithmetic", h1)
+    assert h2 == 2809036171121327430
+
+
+def test_hash_empty_key() -> None:
+    h = esxi_hash(b"", 666)
+    assert h == 8664614747486377173
